requirements.txt
asgiref==3.8.1
boto3==1.37.6
botocore==1.37.6
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==4.2.19
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
django-phonenumber-field==8.0.0
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
jmespath==1.0.1
phonenumbers==8.13.55
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==1.26.20
